refactor: extract EPUB inspection helpers
